requirements.txt
aiogram>=3.4.1
fastapi[all]>=0.110.0
aiosqlite>=0.20.0
Pillow>=10.3.0
httpx>=0.27.0
python-dotenv>=1.0.1
